size_t issues
Former-commit-id: 4213c2d25f49954d010ec25e13549c203a0a2b1b

+++ applications/programs/classify_kerdensom/classify_kerdensom_main.cpp
@@ -247,7 +247,7 @@ public:
         MetaData vectorHeaderIn, vectorHeaderOut, vectorContentOut;
         vectorHeaderIn.read(formatString("vectorHeader@%s",fn_in.c_str()));
         vectorHeaderOut.setColumnFormat(false);
-        int size, vectorSize;
+        size_t size, vectorSize;
         size_t idIn=vectorHeaderIn.firstObject();
         size_t idOut=vectorHeaderOut.addObject();
         vectorHeaderIn.getValue(MDL_XSIZE,size,idIn);
@@ -256,7 +256,7 @@ public:
         vectorHeaderOut.setValue(MDL_YSIZE,size,idOut);
         vectorHeaderIn.getValue(MDL_ZSIZE,size,idIn);
         vectorHeaderOut.setValue(MDL_ZSIZE,size,idOut);
-        vectorHeaderOut.setValue(MDL_COUNT,(size_t)myMap->size(),idOut);
+        vectorHeaderOut.setValue(MDL_COUNT,myMap->size(),idOut);
         vectorHeaderIn.getValue(MDL_CLASSIFICATION_DATA_SIZE,vectorSize,idIn);
         vectorHeaderOut.setValue(MDL_CLASSIFICATION_DATA_SIZE,vectorSize,idOut);
         vectorHeaderOut.write(formatString("vectorHeader@%s_vectors.xmd",fn_root.c_str()),MD_APPEND);
